combine header files of electron related producers with cc files
