+++ RecoHI/Configuration/python/Reconstruction_HI_cff.py
@@ -9,7 +9,7 @@ from RecoHI.HiTracking.HiTracking_cff import *    # two additional steps
 # Egamma
 from RecoHI.HiEgammaAlgos.HiEgamma_cff import *
 from RecoHI.HiEgammaAlgos.HiElectronSequence_cff import *
-ecalDrivenElectronSeeds.SeedConfiguration.SCEtCut = cms.double(15.0)
+ecalDrivenElectronSeeds.SCEtCut = cms.double(15.0)
 ecalDrivenGsfElectrons.preselection.minSCEtBarrel = cms.double(15.0)
 ecalDrivenGsfElectrons.preselection.minSCEtEndcaps = cms.double(15.0)
