+++ RecoParticleFlow/PFTracking/plugins/modules.cc
@@ -1,5 +1,4 @@
 #include "FWCore/Framework/interface/MakerMacros.h"
-#include "RecoParticleFlow/PFTracking/interface/PFElecTkProducer.h"
 #include "RecoParticleFlow/PFTracking/interface/LightPFTrackProducer.h"
 #include "RecoParticleFlow/PFTracking/interface/PFNuclearProducer.h"
 #include "RecoParticleFlow/PFTracking/plugins/PFConversionProducer.h"
@@ -11,7 +10,6 @@
 #include "RecoParticleFlow/PFTracking/plugins/PFDisplacedVertexSelector.cc"
 #include "RecoParticleFlow/PFTracking/plugins/SimVertexSelector.cc"
 
-DEFINE_FWK_MODULE(PFElecTkProducer);
 DEFINE_FWK_MODULE(LightPFTrackProducer);
 DEFINE_FWK_MODULE(PFNuclearProducer);
 DEFINE_FWK_MODULE(PFConversionProducer);
