+++ RecoTracker/TrackProducer/plugins/SealModules.cc
@@ -3,7 +3,6 @@
 #include "RecoTracker/TrackProducer/plugins/DAFTrackProducer.h"
 #include "RecoTracker/TrackProducer/plugins/TrackProducer.h"
 #include "RecoTracker/TrackProducer/plugins/TrackRefitter.h"
-#include "RecoTracker/TrackProducer/plugins/GsfTrackProducer.h"
 #include "RecoTracker/TrackProducer/plugins/GsfTrackRefitter.h"
 #include "RecoTracker/TrackProducer/plugins/ExtraFromSeeds.h"
 
@@ -11,6 +10,5 @@
 DEFINE_FWK_MODULE(DAFTrackProducer);
 DEFINE_FWK_MODULE(TrackProducer);
 DEFINE_FWK_MODULE(TrackRefitter);
-DEFINE_FWK_MODULE(GsfTrackProducer);
 DEFINE_FWK_MODULE(GsfTrackRefitter);
 DEFINE_FWK_MODULE(ExtraFromSeeds);
